Surpress compiler warning
qmodbusrtuserialslave_p.h: In lambda function:
qmodbusrtuserialslave_p.h:242:17: warning: case value ‘10’ not in enumerated type ‘QModbusPdu::FunctionCode’ [-Wswitch]
     case 0x0a: // Poll 484 (not in the official Modbus specification) *1
                      ^
qmodbusrtuserialslave_p.h:243:17: warning: case value ‘14’ not in enumerated type ‘QModbusPdu::FunctionCode’ [-Wswitch]
     case 0x0e: // Poll Controller (not in the official Modbus specification) *1
                                                                    ^

+++ src/serialbus/qmodbusrtuserialslave_p.h
@@ -238,7 +238,7 @@ public:
                 // its last restart, clear counters operation, or power–up.
                 incrementCounter(QModbusServerPrivate::Counter::BusExceptionError);
             } else {
-                switch (req.functionCode()) {
+                switch (quint16(req.functionCode())) {
                 case 0x0a: // Poll 484 (not in the official Modbus specification) *1
                 case 0x0e: // Poll Controller (not in the official Modbus specification) *1
                 case QModbusRequest::GetCommEventCounter: // fall through and bail out
